requirements.txt
ultragpt==7.5.0
pydantic>=2.5.0,<3
python-dotenv>=1.0.1
ultraprint==3.5.0